- fixed multiple issues - made the ui look good - and many other

+++ game/Models/Buildings/People/BDHouse.h
@@ -10,6 +10,6 @@
 
 @interface BDHouse : BDBuilding
 
-@property (nonatomic, assign) int peopleProduced;
+@property (nonatomic, assign) NSInteger peopleProduced;
 
 @end
